requirements.txt
adium-theme-ubuntu==0.3.4
Django==1.9.6
django-braces==1.9.0
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jsonp==1.0.2
drf-nested-routers==0.11.1
Markdown==2.6.6
mysqlclient==1.3.7
Pillow==3.2.0
pkg-resources==0.0.0
pytz==2014.10
sqlparse==0.1.18
unity-lens-photos==1.0
virtualenv==15.0.1
